requirements/developments.txt
asgiref==3.8.1
autopep8==2.3.2
cffi==1.17.1
cryptography==44.0.3
dj-database-url==2.3.0
Django==5.2
django-autoslug==1.9.9
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-silk==5.3.2
django-unfold==0.59.0
djangorestframework==3.16.0
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.1.0
gprof2dot==2025.4.14
inflection==0.5.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
